lstsize now returns size_t

+++ libft/src/lst/ft_lstsize.c
@@ -6,16 +6,16 @@
 /*   By: mvan-wij <mvan-wij@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2020/10/28 12:59:50 by mvan-wij      #+#    #+#                 */
-/*   Updated: 2021/02/24 17:05:54 by mvan-wij      ########   odam.nl         */
+/*   Updated: 2022/01/24 17:31:23 by mvan-wij      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 #include <stdlib.h>
 
-int	ft_lstsize(t_list *lst)
+size_t	ft_lstsize(t_list *lst)
 {
-	int	len;
+	size_t	len;
 
 	if (lst == NULL)
 		return (0);
