+++ libft/src/lst/libft_lst.h
@@ -6,12 +6,14 @@
 /*   By: mvan-wij <mvan-wij@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2021/07/27 14:55:06 by mvan-wij      #+#    #+#                 */
-/*   Updated: 2021/11/01 11:13:26 by mvan-wij      ########   odam.nl         */
+/*   Updated: 2022/01/24 18:09:10 by mvan-wij      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef LIBFT_LST_H
 # define LIBFT_LST_H
+
+# include <stddef.h>
 
 typedef struct s_list
 {
@@ -28,7 +30,7 @@ void	ft_lstiter(t_list *lst, void (*f)(void *));
 t_list	*ft_lstlast(t_list *lst);
 t_list	*ft_lstmap(t_list *lst, void *(*f)(void *), void (*del)(void *));
 t_list	*ft_lstnew(void *content);
-int		ft_lstsize(t_list *lst);
+size_t	ft_lstsize(t_list *lst);
 t_list	*ft_lstitem(t_list *lst, int i);
 void	ft_lstreverse(t_list **lst);
 t_list	*ft_lstnew_front(void *content, t_list **lst);
